reduced server code
