+++ main.go
@@ -1,48 +1,12 @@
 package main
 
 import (
-	"context"
-	"errors"
-	"log"
-	"net/http"
-	"os"
-	"os/signal"
-	"syscall"
-	"time"
-
 	"github.com/echo-webkom/goat/server"
 	"github.com/joho/godotenv"
 )
 
 func main() {
 	godotenv.Load() // Debug
-
 	s := server.New()
-	s.MountHandlers()
-
-	server := &http.Server{
-		Addr:    s.Config.Addr,
-		Handler: s.Router,
-	}
-
-	go func() {
-		log.Println("Listening at", server.Addr)
-		if err := server.ListenAndServe(); !errors.Is(err, http.ErrServerClosed) {
-			log.Fatalf("HTTP server error: %v", err)
-		}
-		log.Println("Stopped serving new connections. Shutting down...")
-	}()
-
-	// Handle interupt (ctrl-c etc) and gracefully shutdown
-	sigChan := make(chan os.Signal, 1)
-	signal.Notify(sigChan, syscall.SIGINT, syscall.SIGTERM)
-	<-sigChan
-
-	shutdownCtx, shutdownRelease := context.WithTimeout(context.Background(), 10*time.Second)
-	defer shutdownRelease()
-
-	if err := server.Shutdown(shutdownCtx); err != nil {
-		log.Fatalf("HTTP shutdown error: %v", err)
-	}
-	log.Println("Graceful shutdown complete.")
+	server.ServeWithShutdown(s)
 }
